시각화, 그래프 작성
